backend/requirements.txt
psycopg2-binary==2.9.6
gunicorn==20.1.0
Django==3.2.3
django-filter==23.5
django-cors-headers==3.13.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-urlshortner==0.0.2
drf_extra_fields==3.7.0
flake8==5.0.4
python-dotenv==1.0.1
djoser==2.2.0
webcolors==1.11.1
Pillow==9.5.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0